Remove some test code.
Refactor modal popup to be loaded on-demand by a loder item.
Move modal popup states and transitions into it's component file, instead of being in BasicPage.
Decouple the way the side window click handler shows the modal popup by using signals and slots, instead of using flags, and giving a reference to the popup item.