Fix bug related to exercise immediate transition
when loading an existing workout that has a completed exercise, not using a delayed animation when click on an exercise set.
Move side window component into a loder, so it's loaded on demand.
Fix issue with content item being enabled when a side window is being closed, thus possibly leading to unwanted mouse click events.
Implement saving of workout data when clicking on back button, and there is data to be saved (at least one exercise set is completed).