Adjust workout query to extract previous workout
data not specific to a day, but rather to a workout template, for a user, thus returning a list of exercises that were previously done, along with the work weight and if they were successfully completed.
The query is sorted in order to get the data in the last attempt, for each specific exercise, which can span different days.