Add Triange QML Component.
Add initial implementation of graph drag handle and bubble.
Add opacity animation for lines, when a line is clicked on and the handle is dragged.
Add possiblity of drawing points on the graph for debug purposes.
Remove old code in graph page.
Add functions to figure out closes exercise point and related data, a given mouse coordinate point.