Add feature to reset plate weights to defaults.
Add feature to sum up plates with the same weight, if there are duplicate plates with the same weight.
Add better support for removal, by allowing 0 in both weight and count.
Fix issue with too many digits being after the dot, when converting a double value to a string in JS.
Fix issue with multiple settings pages being open if clicking multiple times on a cell in a short period of time.