Add skeleton implementation for exercise sets and reps settings page.
Exercise sets and reps are the corresponding chosen set & rep entities for a specific exercises in a template, for a specific user (basically the UI page that you access when on the workout page, and you click the weight / or set & rep component).
Fix SQL bug, which showed incorrect set and rep for a new workout (it used the current workout id if existed, otherwise defaulting to default id_set_and_rep for an exercise, now it uses current workout id if present, falling back to previous workout id, and then falling back to the default id_set_and_rep for an exercise).